+++ packages/common-all/src/store/NoteMetadataStore.ts
@@ -2,12 +2,8 @@ import _ from "lodash";
 import { ERROR_STATUS, ERROR_SEVERITY } from "../constants";
 import { DendronError } from "../error";
 import { NoteFnameDictUtils } from "../noteDictsUtils";
-import {
-  NotePropsMeta,
-  NotePropsByFnameDict,
-  RespV3,
-  FindNoteOpts,
-} from "../types";
+import { NotePropsMeta, NotePropsByFnameDict, RespV3 } from "../types";
+import { FindNoteOpts } from "../types/FindNoteOpts";
 import { cleanName, isNotUndefined } from "../utils";
 import { VaultUtils } from "../vault";
 import { IDataStore } from "./IDataStore";
