+++ packages/common-all/src/types/DWorkspace.ts
@@ -1,0 +1,10 @@
+import { DVault } from "./DVault";
+import { RemoteEndpoint } from "./RemoteEndpoint";
+
+export type DWorkspace = {
+  name: string;
+  vaults: DVault[];
+  remote: RemoteEndpoint;
+};
+
+export type DWorkspaceEntry = Omit<DWorkspace, "name" | "vaults">;
