+++ packages/common-all/src/types/DendronWorkspace.ts
@@ -1,0 +1,8 @@
+import { DVault } from "./DVault";
+import { RemoteEndpoint } from "./RemoteEndpoint";
+
+export type DendronWorkspace = {
+  name: string;
+  vaults: DVault[];
+  remote: RemoteEndpoint;
+};
