+++ packages/common-all/src/types/DendronWorkspaceEntry.ts
@@ -1,0 +1,3 @@
+import { DendronWorkspace } from "./DendronWorkspace";
+
+export type DendronWorkspaceEntry = Omit<DendronWorkspace, "name" | "vaults">;
