+++ packages/common-all/src/types/RemoteEndpoint.ts
@@ -1,0 +1,4 @@
+export type RemoteEndpoint = {
+  type: "git";
+  url: string;
+};
