+++ packages/common-all/src/types/SeedEntry.ts
@@ -1,0 +1,12 @@
+import { SeedSite } from "./seed";
+
+export type SeedEntry = {
+  /**
+   * Specific branch to pull from
+   */
+  branch?: string;
+  /**
+   * When in this seed, what url to use
+   */
+  site?: SeedSite;
+};
