+++ packages/common-all/src/types/configs/workspace/index.ts
@@ -2,5 +2,5 @@ export * from "./graph";
 export * from "./journal";
 export * from "./scratch";
 export * from "./types";
-export * from "./workspace";
+export * from "./DendronWorkspaceConfig";
 export * from "./task";
