+++ packages/common-all/src/types/typesv2.ts
@@ -3,6 +3,8 @@ import { IDendronError } from "../error";
 import {
   DNodeProps,
   DNodeType,
+  DNoteAnchorPositioned,
+  DNoteLoc,
   NoteProps,
   NotePropsMeta,
   Position,
@@ -10,12 +12,12 @@ import {
   SchemaProps,
 } from "./foundation";
 import { DHookDict } from "./hooks";
-import { DVault } from "./workspace";
 import { IntermediateDendronConfig } from "./intermediateConfigs";
 import { VSRange } from "./compat";
 import { Decoration, Diagnostic } from ".";
-import { FindNoteOpts } from "./store";
 import { DendronASTDest, ProcFlavor } from "./unified";
+import { FindNoteOpts } from "./FindNoteOpts";
+import { DVault } from "./DVault";
 
 export type OptionalExceptFor<T, TRequired extends keyof T> = Partial<T> &
   Pick<T, TRequired>;
@@ -32,63 +34,6 @@ export type EngineDeleteOpts = {
 };
 
 // === New
-
-export type DNoteLoc = {
-  fname: string;
-  alias?: string;
-  id?: string;
-  vaultName?: string;
-  anchorHeader?: string;
-};
-
-export type DNoteAnchor =
-  | DNoteBlockAnchor
-  | DNoteHeaderAnchor
-  | DNoteLineAnchor;
-
-/**
- * Anchor without {@link DNoteHeaderAnchor.depth} info
- * @todo see migration [[DNoteAnchorBasic|dendron://dendron.docs/dev.changelog#dnoteanchorbasic]]
- */
-export type DNoteAnchorBasic =
-  | DNoteBlockAnchor
-  | Omit<DNoteHeaderAnchor, "depth">
-  | DNoteLineAnchor;
-
-export type DNoteBlockAnchor = {
-  type: "block";
-  text?: string; //original text for the anchor
-  value: string;
-};
-
-/**
- * This represents a markdown header
- * ```md
- * # H1
- * ```
- */
-export type DNoteHeaderAnchor = {
-  type: "header";
-  text?: string; //original text for the anchor
-  value: string;
-  depth: number;
-};
-
-/** An anchor referring to a specific line in a file. These don't exist inside of files, they are implied by the link containing the anchor.
- *
- * Lines are indexed starting at 1, which is similar to how you refer to specific lines on Github.
- */
-export type DNoteLineAnchor = {
-  type: "line";
-  /** 1-indexed line number. */
-  line: number;
-  value: string;
-};
-
-export type DNoteAnchorPositioned = (DNoteBlockAnchor | DNoteHeaderAnchor) & {
-  line: number;
-  column: number;
-};
 
 export type DLinkType = "wiki" | "refv2" | "hashtag" | "usertag" | "fmtag";
